+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LIBREAUTH_OATH_CODE_INVALID_UTF8` has been renamed `LIBREAUTH_OATH_INVALID_UTF8`.
 - The `output_base_len` parameter in the `libreauth_hotp_cfg` and `libreauth_totp_cfg` structures has been dropped.
 - The PBKDF2 `hash` parameter has been renamed `hmac`.
+- The HashFunction has been moved to a dedicated module and its C-bindings renamed accordingly.
 
 
 ## [0.12.0] - 2019-08-27